+++ mod-audit-server/src/main/java/org/folio/services/management/AuditManager.java
@@ -10,6 +10,7 @@ import org.folio.services.configuration.SettingGroup;
 import org.folio.services.configuration.SettingKey;
 import org.folio.services.inventory.InventoryEventService;
 import org.folio.services.marc.MarcAuditService;
+import org.folio.services.user.UserEventService;
 import org.folio.util.marc.SourceRecordType;
 import org.springframework.stereotype.Service;
 
@@ -22,13 +23,16 @@ public class AuditManager {
   private final InventoryEventService inventoryService;
   private final MarcAuditService marcService;
   private final PartitionService partitionService;
+  private final UserEventService userEventService;
 
   public AuditManager(ConfigurationService configurationService, InventoryEventService inventoryService,
-                      MarcAuditService marcService, PartitionService partitionService) {
+                      MarcAuditService marcService, PartitionService partitionService,
+                      UserEventService userEventService) {
     this.configurationService = configurationService;
     this.inventoryService = inventoryService;
     this.marcService = marcService;
     this.partitionService = partitionService;
+    this.userEventService = userEventService;
   }
 
   public Future<Void> executeDatabaseCleanup(String tenantId) {
@@ -46,6 +50,8 @@ public class AuditManager {
       deleteExpiredRecordsForSettingGroup(tenantId, SettingGroup.AUTHORITY, currentTime,
         (tenant, expireOlderThan) ->
           marcService.expireRecords(tenant, expireOlderThan, SourceRecordType.MARC_AUTHORITY)),
+      deleteExpiredRecordsForSettingGroup(tenantId, SettingGroup.USER, currentTime,
+        userEventService::expireRecords),
       partitionService.cleanUpAndCreateSubPartitions(tenantId)
     ).mapEmpty();
   }